% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colorChooser}
\alias{colorChooser}
\title{Pick a color}
\arguments{
\item{title}{\code{base::character} \cr Window title}

\item{hexrgb}{\code{base::character} \cr Pre-selected color}
}
\value{
A \code{character(1)} of the selected hex value, or NULL if the dialog was cancelled.
}
\description{
Pick a color using a system dialog.
}
