% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{messageBox}
\alias{messageBox}
\title{Display a message}
\arguments{
\item{title}{\code{base::character} \cr Window title}

\item{message}{\code{base::character} \cr A message to display}

\item{diagtype}{\code{base::character} \cr Dialog type (buttons), one of: "ok", "okcancel", "yesno", "yesnocancel"}

\item{icontype}{\code{base::character} \cr Icon displayed on the dialog, one of: "info", "warning", "error", "question"}

\item{value}{\code{base::integer} \cr Default button focus depending on dialog type. 0 for cancel/no, 1 for ok/yes, and 2 for no when diagtype is "yesnocancel"}
}
\value{
A \code{character(1)} of the entered string, or NULL if the dialog was cancelled.
}
\description{
Show a modal dialog with configurable buttons and icon.
}
