% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFileDialog.R
\name{openFileDialog}
\alias{openFileDialog}
\title{Select one or more files}
\usage{
openFileDialog(
  title = "Open File",
  filepath = getwd(),
  filters = "",
  description = "",
  multiple = F
)
}
\arguments{
\item{title}{\code{base::character} Window title}

\item{filepath}{\code{base::character} A default filename and/or path}

\item{filters}{\code{base::character} File extension filter patterns}

\item{description}{\code{base::character} Generic file type descriptor string}

\item{multiple}{\code{base::logical} Allow multiple files to be selected?}
}
\value{
A \code{character()} vector of the selected file(s), or NULL if the dialog was cancelled.
}
\description{
Prompt the user to select one or more files (if allowed) of the given extensions using the standard system dialog.
}
