% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFileDialog.R
\name{saveFileDialog}
\alias{saveFileDialog}
\title{Save a file}
\usage{
saveFileDialog(
  title = "Save File",
  filepath = getwd(),
  filters = "",
  description = ""
)
}
\arguments{
\item{title}{\code{base::character} Window title}

\item{filepath}{\code{base::character} A default filename and/or path}

\item{filters}{\code{base::character} File extension filter patterns}

\item{description}{\code{base::character} Generic file type descriptor string}
}
\value{
A \code{character(1)} of the selected file, or NULL if the dialog was cancelled.
}
\description{
Prompt the user to save a file with a given extension using a system dialog.
}
