% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFolderDialog.R
\name{selectFolderDialog}
\alias{selectFolderDialog}
\title{Choose a folder}
\usage{
selectFolderDialog(title = "Choose Folder", path = getwd())
}
\arguments{
\item{title}{\code{base::character} \cr Window title}

\item{path}{\code{base::character} \cr A default path}
}
\value{
A \code{character(1)} of the selected folder, or NULL if the dialog was cancelled.
}
\description{
Prompt the user to select a folder.
}
